+++ crates/ryo3-jiff/src/round.rs
@@ -113,7 +113,7 @@ impl RoundOptions {
 // DateTimeRound
 // ----------------------------------------------------------------------------
 #[derive(Clone, Copy, Debug)]
-#[pyclass(name = "DateTimeRound", frozen, immutable_type)]
+#[pyclass(name = "DateTimeRound", frozen, immutable_type, skip_from_py_object)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RyDateTimeRound {
     pub(crate) options: RoundOptions,
@@ -220,7 +220,12 @@ impl Display for RyDateTimeRound {
 // SignedDurationRound
 // ----------------------------------------------------------------------------
 #[derive(Clone, Copy, Debug)]
-#[pyclass(name = "SignedDurationRound", frozen, immutable_type)]
+#[pyclass(
+    name = "SignedDurationRound",
+    frozen,
+    immutable_type,
+    skip_from_py_object
+)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RySignedDurationRound {
     options: RoundOptions,
@@ -324,7 +329,7 @@ impl Display for RySignedDurationRound {
 // TimeRound
 // ----------------------------------------------------------------------------
 #[derive(Clone, Copy, Debug)]
-#[pyclass(name = "TimeRound", frozen, immutable_type)]
+#[pyclass(name = "TimeRound", frozen, immutable_type, skip_from_py_object)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RyTimeRound {
     options: RoundOptions,
@@ -428,7 +433,7 @@ impl Display for RyTimeRound {
 // TimestampRound
 // ----------------------------------------------------------------------------
 #[derive(Clone, Copy, Debug)]
-#[pyclass(name = "TimestampRound", frozen, immutable_type)]
+#[pyclass(name = "TimestampRound", frozen, immutable_type, skip_from_py_object)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RyTimestampRound {
     options: RoundOptions,
@@ -533,7 +538,12 @@ impl Display for RyTimestampRound {
 // ZonedDateTimeRound
 // ---------------------------------------------------------------------------
 #[derive(Clone, Copy, Debug)]
-#[pyclass(name = "ZonedDateTimeRound", frozen, immutable_type)]
+#[pyclass(
+    name = "ZonedDateTimeRound",
+    frozen,
+    immutable_type,
+    skip_from_py_object
+)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RyZonedDateTimeRound {
     options: RoundOptions,
@@ -639,7 +649,7 @@ impl Display for RyZonedDateTimeRound {
 // OffsetRound
 // ----------------------------------------------------------------------------
 #[derive(Clone, Copy, Debug)]
-#[pyclass(name = "OffsetRound", frozen, immutable_type)]
+#[pyclass(name = "OffsetRound", frozen, immutable_type, skip_from_py_object)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RyOffsetRound {
     options: RoundOptions,