+++ crates/ryo3-jiff/src/ry_zoned.rs
@@ -29,7 +29,7 @@ use std::hash::{DefaultHasher, Hash, Hasher};
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
 #[derive(Debug, Clone, PartialEq)]
-#[pyclass(name = "ZonedDateTime", frozen, immutable_type)]
+#[pyclass(name = "ZonedDateTime", frozen, immutable_type, from_py_object)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct RyZoned(pub(crate) Zoned);
 
@@ -571,8 +571,8 @@ impl RyZoned {
     fn replace(
         &self,
         obj: Option<Bound<'_, PyAny>>,
-        date: Option<RyDate>,
-        time: Option<RyTime>,
+        date: Option<&RyDate>,
+        time: Option<&RyTime>,
         year: Option<i16>,
         era_year: Option<(i16, JiffEra)>,
         month: Option<i8>,
@@ -586,7 +586,7 @@ impl RyZoned {
         microsecond: Option<i16>,
         nanosecond: Option<i16>,
         subsec_nanosecond: Option<i32>,
-        offset: Option<RyOffset>,
+        offset: Option<&RyOffset>,
         offset_conflict: Option<JiffTzOffsetConflict>,
         disambiguation: Option<JiffTzDisambiguation>,
     ) -> PyResult<Self> {
@@ -594,17 +594,14 @@ impl RyZoned {
         let mut builder = self.0.with();
         if let Some(obj) = obj {
             // if obj is a Zoned, use it as the base
-            if let Ok(zoned) = obj.cast::<RyDate>() {
+            if let Ok(zoned) = obj.cast_exact::<RyDate>() {
                 // if obj is a Zoned, use it as the base
-                let date = zoned.extract::<RyDate>()?;
-                builder = builder.date(date.0);
-            } else if let Ok(time) = obj.cast::<RyTime>() {
+                builder = builder.date(zoned.get().0);
+            } else if let Ok(time) = obj.cast_exact::<RyTime>() {
                 // if obj is a Time, use it as the base
-                let time = time.extract::<RyTime>()?;
-                builder = builder.time(time.0);
-            } else if let Ok(date) = obj.cast::<RyOffset>() {
-                let offset = date.extract::<RyOffset>()?;
-                builder = builder.offset(offset.0);
+                builder = builder.time(time.get().0);
+            } else if let Ok(offset) = obj.cast_exact::<RyOffset>() {
+                builder = builder.offset(offset.get().0);
             } else {
                 return py_type_err!("obj must be a Date, Time or Offset; given: {obj}",);
             }