+++ crates/ryo3-tokio/src/fs/async_file_read_stream.rs
@@ -114,7 +114,12 @@ impl AsyncFileReadStreamWrapper {
 
 impl AsyncFileReadStreamWrapper {}
 
-#[pyclass(name = "AsyncFileReadStream", frozen, immutable_type)]
+#[pyclass(
+    name = "AsyncFileReadStream",
+    frozen,
+    immutable_type,
+    skip_from_py_object
+)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 pub struct PyAsyncFileReadStream {
     options: Arc<PyFileReadStreamOptions>,